pyproject.toml
[project]
name = "devboard-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.44.0",
]
